% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g1.R
\name{calc_g1}
\alias{calc_g1}
\title{Calculates the autocorrelation function value (g1(t)) from measured correlation time and related intensity autocorrelation (g2(t))}
\usage{
calc_g1(d, n)
}
\arguments{
\item{d}{A tibble consisting of measured correlation time and related intensity autocorrelation (g2(t)) value}

\item{n}{An integer}
}
\value{
A tibble consisting of correlation time, calculated g1(t) and scaled g1(t)
}
\description{
Calculates the autocorrelation function value (g1(t)) from measured correlation time and related intensity autocorrelation (g2(t))
}
\examples{
e <- calc_g1(d, 30)
}
