% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{calc_GDprime}
\alias{calc_GDprime}
\title{This function is used to calculate the loss modulus}
\usage{
calc_GDprime(G, slope)
}
\arguments{
\item{G}{A vector (viscoelastic modulus)}

\item{slope}{A vector}
}
\value{
A vector
}
\description{
This function is used to calculate the loss modulus
}
\examples{
Loss <- calc_GDprime(G, slope)
}
