% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{calc_Gprime}
\alias{calc_Gprime}
\title{This function is used to calculate the storage modulus}
\usage{
calc_Gprime(G, slope)
}
\arguments{
\item{G}{A vector (viscoelastic modulus)}

\item{slope}{A vector}
}
\value{
A vector
}
\description{
This function is used to calculate the storage modulus
}
\examples{
Storage <- calc_Gprime(G, slope)
}
