% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{calc_modulus}
\alias{calc_modulus}
\title{This function produces a tibble consisting of frequency, the storage and loss modulus}
\usage{
calc_modulus(d, temp, radius)
}
\arguments{
\item{temp}{The temperature as Celsius}

\item{radius}{A number, relating to sphere size (eg 5e-7)}

\item{G}{A tibble consisting of correlation time and related mean square displacement}
}
\value{
A tibble consisting of frequency and the related storage and loss modulus
}
\description{
This function produces a tibble consisting of frequency, the storage and loss modulus
}
\examples{
e <- calc_modulus(d, 20, 5e-7)
}
