% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{calc_slope}
\alias{calc_slope}
\title{This function is used internally for calculating the viscoelastic modulus.}
\usage{
calc_slope(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector (same length as x)}
}
\value{
A vector
}
\description{
This function is used internally for calculating the viscoelastic modulus.
}
\examples{
calc_slope(time, msd)
}
