% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{This function is used internally for calculating modulus, converting temperature in Celsius to absolute (Kelvin)}
\usage{
celsius_to_kelvin(temp)
}
\arguments{
\item{temp}{A numnber}
}
\value{
A number
}
\description{
This function is used internally for calculating modulus, converting temperature in Celsius to absolute (Kelvin)
}
\examples{
celsius_to_kelvin(20)
}
