% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dws}
\alias{dws}
\title{Set of related DWS data.}
\format{A data frame with 288 rows and 2 variables:
\describe{
  \item{time}{correlation time, in seconds}
  \item{g2}{measured intensity autocorrelation values, dimensionless}
  ...
}}
\usage{
dws
}
\description{
A dataset containing measured correlation times with related intensity autocorrelation values.
}
\keyword{datasets}
