% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{FindX}
\alias{FindX}
\title{Used internally for calculating mean square displacement for transmission geometry}
\usage{
FindX(y)
}
\description{
Used internally for calculating mean square displacement for transmission geometry
}
