% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{form_freq}
\alias{form_freq}
\title{This function is used to calculate the frequency as inverse of time}
\usage{
form_freq(time)
}
\arguments{
\item{time}{A vector}
}
\value{
A vector
}
\description{
This function is used to calculate the frequency as inverse of time
}
\examples{
form_freq(time)
}
