% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g1.R
\name{form_g1}
\alias{form_g1}
\title{Calculates the autocorrelation function value (g1(t)) from measured correlation time and related intensity autocorrelation (g2(t))}
\usage{
form_g1(t_g2, num_points = NULL)
}
\arguments{
\item{t_g2}{A tibble consisting of measured correlation time and related intensity autocorrelation (g2(t)) value}

\item{num_points}{(optional) number of data points for scaling, default = 30}
}
\value{
A tibble consisting of correlation time, calculated g1(t) and scaled g1(t)
}
\description{
Calculates the autocorrelation function value (g1(t)) from measured correlation time and related intensity autocorrelation (g2(t))
}
