% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{form_slope}
\alias{form_slope}
\title{This function is used internally for calculating the viscoelastic modulus.}
\usage{
form_slope(time, msd)
}
\arguments{
\item{time}{A vector}

\item{msd}{A vector (same length as x)}
}
\value{
A vector
}
\description{
This function is used internally for calculating the viscoelastic modulus.
}
\examples{
slope(time, msd)
}
