% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g1.R
\name{g1_spline}
\alias{g1_spline}
\title{Interpolate the scaled g1 vs time function and return fitted values}
\usage{
g1_spline(d)
}
\arguments{
\item{d}{A tibble consisting of time and scaled g1(t) values}
}
\value{
A numeric vector.
}
\description{
Interpolate the scaled g1 vs time function and return fitted values
}
\examples{
pred <- g1_spline(d)
}
