% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{loss_mod}
\alias{loss_mod}
\title{This function is used to calculate the loss modulus}
\usage{
loss_mod(visco_mod, slope)
}
\arguments{
\item{visco_mod}{A vector (viscoelastic modulus)}

\item{slope}{A vector}
}
\value{
A vector
}
\description{
This function is used to calculate the loss modulus
}
