% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g1.R
\name{range_scale}
\alias{range_scale}
\title{Range scale a vector using the mean of first n elements (num_points) as the maximum value}
\usage{
range_scale(vec, num_points)
}
\arguments{
\item{vec}{A vector}

\item{num_points}{An integer}
}
\value{
A numeric vector
}
\description{
Range scale a vector using the mean of first n elements (num_points) as the maximum value
}
\examples{
range_scale(observed, 30)
}
