% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{visco_mod}
\alias{visco_mod}
\title{This function is used to calculate the viscoelastic modulus.}
\usage{
visco_mod(temp, radius, msd, slope)
}
\arguments{
\item{temp}{A number (temperature, in Celsius)}

\item{radius}{A number}

\item{msd}{A vector}

\item{slope}{A vector}
}
\value{
A vector
}
\description{
This function is used to calculate the viscoelastic modulus.
}
