% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{calc_freq}
\alias{calc_freq}
\title{This function is used to calculate the frequency as inverse of time (time_c)}
\usage{
calc_freq(time_c)
}
\arguments{
\item{time_c}{A vector}
}
\value{
A vector
}
\description{
This function is used to calculate the frequency as inverse of time (time_c)
}
\examples{
calc_freq(time_c)
}
