% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{calc_modulus}
\alias{calc_modulus}
\title{This function produces a tibble consisting of frequency, the storage and loss modulus}
\usage{
calc_modulus(d, temp = NULL, radius = NULL)
}
\arguments{
\item{d}{A tibble consisting of correlation time and related mean square displacement}

\item{temp}{(optional) temperature in Celsius, default = 20}

\item{radius}{(optional) particle radius size, default = 5e-7}
}
\value{
A tibble consisting of frequency with related storage and loss modulii
}
\description{
This function produces a tibble consisting of frequency, the storage and loss modulus
}
\examples{
e <- calc_modulus(d, 20, 5e-7)
}
