% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{form_msd}
\alias{form_msd}
\title{Calculate the mean square displacement}
\usage{
form_msd(g1)
}
\arguments{
\item{g1}{A tibble consisting of correlation time, observed and scaled g1(t) values}
}
\value{
A tibble consisting of correlation time and related mean square displacement
}
\description{
Calculate the mean square displacement
}
\examples{
msd <- form_msd(g1)
}
