% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{msd_g1_diff}
\alias{msd_g1_diff}
\title{Used internally for calculating mean square displacement for transmission geometry}
\usage{
msd_g1_diff(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
A function localised in scope
}
\description{
Used internally for calculating mean square displacement for transmission geometry
}
