% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{plot_modulus}
\alias{plot_modulus}
\title{Plots storage and loss modulus against the measured frequency}
\usage{
plot_modulus(mod_t, y_threshold = 1e-06)
}
\arguments{
\item{mod_t}{A tibble consisting of frequency and the related storage and loss modulusA}

\item{y_threshold}{(optional) Minimum 'y' axis plot value, default = 10^-6}
}
\description{
Plots storage and loss modulus against the measured frequency
}
