% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{plot_modulus}
\alias{plot_modulus}
\title{Plots storage and loss modulus against the measured frequency}
\usage{
plot_modulus(g)
}
\arguments{
\item{g}{A tibble consisting of frequency and the related storage and loss modulusA}
}
\description{
Plots storage and loss modulus against the measured frequency
}
\examples{
plot_modulus(g)
}
