% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{plot_msd}
\alias{plot_msd}
\title{Plots the mean square displacement against the correlation time}
\usage{
plot_msd(g1_msd)
}
\arguments{
\item{g1_msd}{A tibble consisting of correlation time and related mean square displacement}
}
\description{
Plots the mean square displacement against the correlation time
}
