% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g1.R
\name{range_scale}
\alias{range_scale}
\title{Range scale a vector using the mean of first n elements as the maximum value}
\usage{
range_scale(y, n)
}
\arguments{
\item{y}{A vector}

\item{n}{An integer}
}
\value{
A numeric vector
}
\description{
Range scale a vector using the mean of first n elements as the maximum value
}
\examples{
range_scale(Observed, 30)
}
